requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==5.1.0
httpx==0.26.0

# Scheduling
apscheduler==3.10.4

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Utilities
python-dateutil==2.8.2
openpyxl==3.1.5
